requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
inflection==0.5.1
install==1.3.5
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
psycopg2-binary==2.8.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1